# shared/openapi.yaml
openapi: 3.0.3
info:
  title: Local AI Data Butler MVP API
  version: 0.1.0
  description: >
    First-version local API contract for the Data Butler MVP. This contract
    covers read-only scanning, local asset indexing, exact duplicate groups,
    file-intelligence overview, explainable suggestions, operation preview,
    quarantine execution, receipts, undo, and mock provider analysis.
    It does not expose permanent delete, real cloud upload, WeChat chat
    database parsing, face recognition, payment, or desktop-shell APIs.
servers:
  - url: http://127.0.0.1:8765
    description: Local Core service
...
paths:
...
  /scan:
    post:
      tags: [scan]
      summary: Start a read-only scan for authorized local directories
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "202":
          description: Scan session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSession"
        default:
          $ref: "#/components/responses/Error"
  /scan/{scan_session_id}:
    get:
      tags: [scan]
      summary: Get scan session state
      operationId: getScanSession
      parameters:
        - $ref: "#/components/parameters/ScanSessionId"
      responses:
        "200":
          description: Scan session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSession"
        default:
          $ref: "#/components/responses/Error"
  /scan/{scan_session_id}/stream:
    get:
      tags: [scan]
      summary: Stream scan progress events
      description: >
        Server-Sent Events endpoint. Each event uses the SSE format
        `event: scan_event` and `data: {json}`. The JSON payload MUST match
        the ScanEvent schema. This endpoint is read-only and MUST NOT change
        files.
      operationId: streamScanEvents
      parameters:
        - $ref: "#/components/parameters/ScanSessionId"
      responses:
        "200":
          description: SSE stream of ScanEvent payloads
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                progress:
                  summary: Scan progress event
                  value: |
                    event: scan_event
                    data: {"event_id":"evt_001","scan_session_id":"scan_001","type":"progress","occurred_at":"2026-06-17T10:00:00Z","current_path":"C:\\Users\\me\\Downloads","files_seen":120,"files_indexed":118,"files_skipped":2}
        default:
          $ref: "#/components/responses/Error"
  /scan/{scan_session_id}/cancel:
    post:
      tags: [scan]
      summary: Cancel a running scan
      operationId: cancelScan
      parameters:
        - $ref: "#/components/parameters/ScanSessionId"
      responses:
        "200":
          description: Updated scan session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanSession"
        default:
          $ref: "#/components/responses/Error"
  /assets:
    get:
      tags: [assets]
      summary: List indexed assets
      operationId: listAssets
      parameters:
        - $ref: "#/components/parameters/ScanSessionIdQuery"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/AssetCategory"
        - name: source_hint
          in: query
          schema:
            $ref: "#/components/schemas/SourceHint"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AssetStatus"
        - name: min_size_bytes
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Asset list
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Asset"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        default:
          $ref: "#/components/responses/Error"
  /assets/{asset_id}:
    get:
      tags: [assets]
      summary: Get asset detail
      operationId: getAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                type: object
                required: [asset, metadata]
                properties:
                  asset:
                    $ref: "#/components/schemas/Asset"
                  metadata:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssetMetadata"
        default:
          $ref: "#/components/responses/Error"
  /fileintel/overview:
    get:
      tags: [fileintel]
      summary: Get local file-intelligence overview
      operationId: getFileIntelOverview
      parameters:
        - $ref: "#/components/parameters/ScanSessionIdQuery"
      responses:
        "200":
          description: File intelligence overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileIntelOverview"
        default:
          $ref: "#/components/responses/Error"
  /dups:
    get:
      tags: [dups]
      summary: List exact duplicate groups
      operationId: listDuplicateGroups
      parameters:
        - $ref: "#/components/parameters/ScanSessionIdQuery"
        - name: min_reclaimable_bytes
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Duplicate groups
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/DuplicateGroup"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        default:
          $ref: "#/components/responses/Error"
  /suggestions:
    get:
      tags: [suggestions]
      summary: List explainable suggestions
      operationId: listSuggestions
      parameters:
        - $ref: "#/components/parameters/ScanSessionIdQuery"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/SuggestionCategory"
        - name: risk_level
          in: query
          schema:
            $ref: "#/components/schemas/RiskLevel"
        - name: default_selected
          in: query
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Suggestion"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        default:
          $ref: "#/components/responses/Error"
  /operations/preview:
    post:
      tags: [operations]
      summary: Create an operation plan and preview
      description: >
        Frontend MUST request a preview before execution. The request uses
        asset_ids or suggestion_ids, never filesystem paths as executable
        authority.
      operationId: createOperationPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationPreviewRequest"
      responses:
        "200":
          description: Operation plan preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationPlan"
        default:
          $ref: "#/components/responses/Error"
  /operations/execute:
    post:
      tags: [operations]
      summary: Execute a confirmed operation plan
      description: >
        Execution MUST be based on operation_plan_id and confirm=true.
        P0 only supports reversible operations such as quarantine or local
        metadata-only ignore. Permanent deletion is not part of this contract.
      operationId: executeOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationExecuteRequest"
      responses:
        "200":
          description: Operation receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationReceipt"
        default:
          $ref: "#/components/responses/Error"
  /operations:
    get:
      tags: [operations]
      summary: List operation receipts
      operationId: listOperations
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OperationStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Operation receipts
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OperationReceipt"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        default:
          $ref: "#/components/responses/Error"
  /operations/{operation_id}:
    get:
      tags: [operations]
      summary: Get operation receipt detail
      operationId: getOperation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationReceipt"
        default:
          $ref: "#/components/responses/Error"
  /operations/{operation_id}/undo:
    post:
      tags: [operations]
      summary: Undo a reversible operation
      operationId: undoOperation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperationUndoRequest"
      responses:
        "200":
          description: Updated operation receipt after undo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationReceipt"
        default:
          $ref: "#/components/responses/Error"
  /quarantine:
    get:
      tags: [quarantine]
      summary: List quarantine items
      operationId: listQuarantineItems
      parameters:
        - name: operation_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/QuarantineStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Quarantine items
          content:
            application/json:
              schema:
                type: object
                required: [items, page]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuarantineItem"
                  page:
                    $ref: "#/components/schemas/PageInfo"
        default:
          $ref: "#/components/responses/Error"
  /providers:
    get:
      tags: [providers]
      summary: List provider capabilities
      description: >
        P0 providers are local-only. Any mock provider listed here MUST NOT
        imply real upload or cloud processing capability.
      operationId: listProviders
      responses:
        "200":
          description: Provider capabilities
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProviderInfo"
        default:
          $ref: "#/components/responses/Error"
  /providers/mock/analyze:
    post:
      tags: [providers]
      summary: Run local mock analysis
      description: >
        Stub endpoint for frontend integration tests. It accepts asset IDs or
        aggregate stats and returns mock analysis. It MUST NOT upload files,
        read remote services, or represent real cloud enhancement capability.
      operationId: runMockProviderAnalyze
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MockAnalyzeRequest"
      responses:
        "200":
          description: Mock provider analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockAnalyzeResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/inventory/scans:
    post:
      tags: [inventory]
      summary: Run the fixed metadata-only inventory fixture
      operationId: createInventoryScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryScanCreateRequest"
      responses:
        "201":
          description: Completed or partial fixture inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryScanResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/inventory/scans/{scan_id}:
    get:
      tags: [inventory]
      summary: Get fixture inventory status
      operationId: getInventoryScan
      parameters:
        - $ref: "#/components/parameters/InventoryScanId"
      responses:
        "200":
          description: Fixture inventory status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryScanResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/inventory/snapshots:
    post:
      tags: [inventory]
      summary: Materialize the persisted inventory snapshot
      operationId: createInventorySnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventorySnapshotCreateRequest"
      responses:
        "201":
          description: Canonical report snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSnapshotV1"
        default:
          $ref: "#/components/responses/Error"
  /v1/inventory/snapshots/{snapshot_id}:
    get:
      tags: [inventory]
      summary: Get a persisted report snapshot
      operationId: getInventorySnapshot
      parameters:
        - $ref: "#/components/parameters/InventorySnapshotId"
      responses:
        "200":
          description: Canonical report snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSnapshotV1"
        default:
          $ref: "#/components/responses/Error"
...
  /v1/inventory/analyses/{analysis_id}:
    get:
      tags: [inventory]
      summary: Get a persisted analysis and receipt
      operationId: getInventoryAnalysis
      parameters:
        - $ref: "#/components/parameters/InventoryAnalysisId"
      responses:
        "200":
          description: Persisted analysis and receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryAnalysisResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    ScanSessionId:
      name: scan_session_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ScanRequest:
      type: object
      required: [paths]
      properties:
        paths:
          type: array
          minItems: 1
          items:
            type: string
          description: User-authorized local directories. Paths are only scan scope, not operation authority.
        options:
          type: object
          properties:
            include_hidden:
              type: boolean
              default: false
            follow_symlinks:
              type: boolean
              enum: [false]
              default: false
              description: P0 never follows symbolic links or junctions.
            compute_hash:
              type: boolean
              default: true
            source_hint_override:
              $ref: "#/components/schemas/SourceHint"
...
    AssetCategory:
      type: string
      enum: [photo, video, archive, installer, document, pdf, audio, other]
...